+++ app/Http/Controllers/IndexController.php
@@ -7,6 +7,7 @@ use GuzzleHttp\Client;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Support\Facades\Config;
+use League\CommonMark\CommonMarkConverter;
 
 class IndexController extends BaseController
 {
@@ -619,6 +620,114 @@ class IndexController extends BaseController
                     'resource_name' => Config::get('web.config.api_resource_name'),
                     'months' => $months,
                     'year' => $year_identifier
+                ]
+            );
+        }
+    }
+
+    public function versionHistory()
+    {
+        $converter = new CommonMarkConverter();
+        $html = $converter->convertToHtml(file_get_contents('../CHANGELOG.md'));
+
+        return view(
+            'version-history',
+            [
+                'display_nav_options' => $this->display_nav_options,
+                'nav_active' => 'version-history',
+                'version_history' => $html
+            ]
+        );
+    }
+
+    public function expenses(Request $request)
+    {
+        $expenses = null;
+
+        $client = new Client([
+            'base_uri' => Config::get('web.config.api_base_url'),
+            'headers' => [
+                'Accept' => 'application/json',
+                'Content-Type' => 'application/json',
+            ],
+        ]);
+
+        $allowed = ['year', 'month', 'category', 'sub_category'];
+        $request_parameters = [];
+        $filtering = [];
+        $parameters = $request->all();
+
+        foreach ($parameters as $parameter => $value) {
+            if (
+                in_array($parameter, $allowed) === true &&
+                $value !== null
+            ) {
+                switch ($parameter) {
+                    case 'year':
+                        $filtering[] = 'Year: ' . $value;
+                        break;
+                    case 'month':
+                        $filtering[] = 'Month: ' . date("F", mktime(0, 0, 0, $value, 10));
+                        break;
+                    case 'category':
+                        try {
+                            $response = $client->get(
+                                Config::get('web.config.api_uri_category') .
+                                '/' . $value
+                            );
+
+                            if ($response->getStatusCode() === 200) {
+                                $filtering[] .= 'Category: ' . json_decode($response->getBody(), true)['name'];
+                            }
+                        } catch (ClientException $e) {
+                            // Do nothing for now
+                        }
+                        break;
+                    case 'sub_category':
+                        try {
+                            $response = $client->get(
+                                Config::get('web.config.api_uri_category') .
+                                '/' . $request_parameters['category'] . '/sub_categories/' . $value
+                            );
+
+                            if ($response->getStatusCode() === 200) {
+                                $filtering[] .= 'Sub category: ' . json_decode($response->getBody(), true)['name'];
+                            }
+                        } catch (ClientException $e) {
+                            // Do nothing for now
+                        }
+                        break;
+                    default:
+                        break;
+                }
+
+                $request_parameters[$parameter] = $value;
+            }
+        }
+
+        $uri = Config::get('web.config.api_uri_items') . '?limit=5';
+        foreach ($request_parameters as $parameter => $value) {
+            $uri .= '&' . $parameter . '=' . $value;
+        }
+        try {
+            $response = $client->get($uri);
+
+            if ($response->getStatusCode() === 200) {
+                $expenses = json_decode($response->getBody(), true);
+            }
+        } catch (ClientException $e) {
+            return redirect()->action('IndexController@index');
+        }
+
+        if ($expenses !== null) {
+            return view(
+                'expenses',
+                [
+                    'display_nav_options' => $this->display_nav_options,
+                    'nav_active' => $this->nav_active,
+                    'resource_name' => Config::get('web.config.api_resource_name'),
+                    'expenses' => $expenses,
+                    'filtering' => implode(', ', $filtering)
                 ]
             );
         }