+++ resources/views/version-history.blade.php
@@ -1,0 +1,7 @@
+@extends('layouts.app')
+
+@section('content')
+    <div class="col-12 col-sm-12 col-md-8 mt-2 mt-lg-2">
+        {!! $version_history !!}
+    </div>
+@endsection