+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-# Changelog
+# Version history
 
 Full changelog for the costs to expect web app.
+
+## 2018-09-30 - v1.02.0
+
+* Added a version history within the app, parses CHANGELOG.
+* Added link to personal site.
+* Summary tables updated, all now include a link to a filtered list.
+* Added an /expenses view, allows filtering by year, month, category and sub category, no UI yet for filtering the list or pagination.
 
 ## 2018-09-21 - v1.01.0